+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to Stability Matrix will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
+## v2.4.2
+### Added
+- Added Japanese UI language option, thanks to kgmkm_mkgm for the translation
+- Language selection available in Settings, and defaults to system language if supported
+
 ## v2.4.1
 ### Fixed
 - Fixed deleting checkpoints not updating the visual grid until the page is refreshed